c# project done and added arduino project
c# project is done, now I just need to check for bugs and test it with
the device for several possible scenarios. Added the arduino project
folder. The config part of the arduino is done know just need to check
it for bugs as well and finish the part of the working loop.